Make case insensitive and add invalid input option
JS code was moved from HTML file to linked JS file. User input is
automatically converted to all lowercase to match computer options. An
additional else if / else statement was added so that invalid entries are
not accepted. For loop was changed to a while loop based on score amount
to accomodate the invalid entry change.